requirements.txt
Django
matplotlib
# Databases
redis
fastapi
uvicorn
pandas
mysql-connector-python
